Refactor `core` module, to break `core_status` into a new source file.
