+++ probe-rs/src/core.rs
@@ -4,18 +4,14 @@ pub use probe_rs_target::{Architecture, CoreAccessOptions};
 use std::time::Duration;
 
 pub mod core_state;
+pub mod core_status;
+pub mod memory_mapped_registers;
 pub mod registers;
 
 pub use core_state::*;
+pub use core_status::*;
+pub use memory_mapped_registers::MemoryMappedRegister;
 pub use registers::*;
-
-/// A memory mapped register, for instance ARM debug registers (DHCSR, etc).
-pub trait MemoryMappedRegister: Clone + From<u32> + Into<u32> + Sized + std::fmt::Debug {
-    /// The register's address in the target memory.
-    const ADDRESS: u64;
-    /// The register's name.
-    const NAME: &'static str;
-}
 
 /// An struct for storing the current state of a core.
 #[derive(Debug, Clone)]
@@ -481,70 +477,4 @@ impl BreakpointId {
     pub fn new(id: usize) -> Self {
         BreakpointId(id)
     }
-}
-
-/// The status of the core.
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum CoreStatus {
-    /// The core is currently running.
-    Running,
-    /// The core is currently halted. This also specifies the reason as a payload.
-    Halted(HaltReason),
-    /// This is a Cortex-M specific status, and will not be set or handled by RISCV code.
-    LockedUp,
-    /// The core is currently sleeping.
-    Sleeping,
-    /// The core state is currently unknown. This is always the case when the core is first created.
-    Unknown,
-}
-
-impl CoreStatus {
-    /// Returns `true` if the core is currently halted.
-    pub fn is_halted(&self) -> bool {
-        matches!(self, CoreStatus::Halted(_))
-    }
-
-    /// Returns `true` if the core is currently running.
-    pub fn is_running(&self) -> bool {
-        self == &Self::Running
-    }
-}
-
-/// When the core halts due to a breakpoint request, some architectures will allow us to distinguish between a software and hardware breakpoint.
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum BreakpointCause {
-    /// We encountered a hardware breakpoint.
-    Hardware,
-    /// We encountered a software breakpoint instruction.
-    Software,
-    /// We were not able to distinguish if this was a hardware or software breakpoint.
-    Unknown,
-}
-
-/// The reason why a core was halted.
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum HaltReason {
-    /// Multiple reasons for a halt.
-    ///
-    /// This can happen for example when a single instruction
-    /// step ends up on a breakpoint, after which both breakpoint and step / request
-    /// are set.
-    Multiple,
-    /// Core halted due to a breakpoint. The cause is `Unknown` if we cannot distinguish between a hardware and software breakpoint.
-    Breakpoint(BreakpointCause),
-    /// Core halted due to an exception, e.g. an
-    /// an interrupt.
-    Exception,
-    /// Core halted due to a data watchpoint
-    Watchpoint,
-    /// Core halted after single step
-    Step,
-    /// Core halted because of a debugger request
-    Request,
-    /// External halt request
-    External,
-    /// Unknown reason for halt.
-    ///
-    /// This can happen for example when the core is already halted when we connect.
-    Unknown,
 }
